[#140372703] Show thumbnail on map (#12)
New Image component handles multiple image sizes that come back from getPhotos() - due to the Photo Map needing a very small thumbnail and home needing bigger.

We're having Images component now handle passing a specific size down to the Image component (200) and the Image component now displays photos (as usual) but with the size necessary for the home page. Now the home page displays slightly smaller images than the 440 we were using - in order to be suitable for a grid display.